# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tabaq | طبق - Food & Dining Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{restaurantId}/menus:
    get:
      operationId: getRestaurantMenus
      tags: [menus]
      summary: Get menus for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant menus
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Menu"
    post:
      operationId: createMenu
      tags: [menus]
      summary: Create a menu for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuRequest"
      responses:
        "201":
          description: Menu created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"

  # ─── Dishes ───────────────────────────────────────────────────────────────
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List available offers
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: integer
        - name: cityId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferListResponse"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create an offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferRequest"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{offerId}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get offer details
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"

  # ─── Vouchers ─────────────────────────────────────────────────────────────
  /vouchers:
    get:
      operationId: listVouchers
      tags: [vouchers]
      summary: List vouchers (user's owned vouchers)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, used, expired]
      responses:
        "200":
          description: List of vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"
    post:
      operationId: purchaseVoucher
      tags: [vouchers]
      summary: Purchase a voucher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseVoucherRequest"
      responses:
        "201":
          description: Voucher purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"
  /vouchers/{voucherId}:
    get:
      operationId: getVoucher
      tags: [vouchers]
      summary: Get voucher details
      parameters:
        - name: voucherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voucher details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"
  /vouchers/{voucherId}/gift:
    post:
      operationId: giftVoucher
      tags: [vouchers]
      summary: Gift a voucher to someone
      parameters:
        - name: voucherId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftVoucherRequest"
      responses:
        "200":
          description: Voucher gifted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"

  # ─── Reviews ──────────────────────────────────────────────────────────────
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List dining events
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: integer
        - name: cityId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a dining event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
components:
  schemas:
...
    Menu:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        type:
          type: string
          enum: [food, drinks, desserts, set_menu, buffet]
        isActive:
          type: boolean
        sections:
          type: array
          items:
            $ref: "#/components/schemas/MenuSection"
      required:
        - id
        - restaurantId
        - nameEn
        - nameAr
        - type
        - isActive
        - sections
    MenuSection:
      type: object
      properties:
        id:
          type: integer
        menuId:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        displayOrder:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/Dish"
      required:
        - id
        - menuId
        - nameEn
        - nameAr
        - displayOrder
        - items
    CreateMenuRequest:
      type: object
      properties:
        nameEn:
          type: string
        nameAr:
          type: string
        type:
          type: string
          enum: [food, drinks, desserts, set_menu, buffet]
      required:
        - nameEn
        - nameAr
        - type

    # ─── Dishes ───────────────────────────────────────────────────────────
    Dish:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        descriptionEn:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        currency:
          type: string
        imageUrl:
          type: string
        avgRating:
          type: number
        reviewCount:
          type: integer
        popularityScore:
          type: number
        isAvailable:
          type: boolean
        isHalal:
          type: boolean
        isVegetarian:
          type: boolean
        isVegan:
          type: boolean
        isGlutenFree:
          type: boolean
        calories:
          type: integer
      required:
        - id
        - restaurantId
        - nameEn
        - nameAr
        - avgRating
        - reviewCount
        - popularityScore
        - isAvailable
...
    Offer:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        restaurantNameEn:
          type: string
        restaurantNameAr:
          type: string
        restaurantCoverImageUrl:
          type: string
        titleEn:
          type: string
        titleAr:
          type: string
        descriptionEn:
          type: string
        descriptionAr:
          type: string
        discountPercent:
          type: number
        originalPrice:
          type: number
        discountedPrice:
          type: number
        currency:
          type: string
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        totalCapacity:
          type: integer
        remainingCapacity:
          type: integer
        isActive:
          type: boolean
        imageUrl:
          type: string
      required:
        - id
        - restaurantId
        - restaurantNameEn
        - restaurantNameAr
        - titleEn
        - titleAr
        - isActive
    OfferListResponse:
      type: object
      properties:
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required:
        - offers
        - total
        - offset
        - limit
    CreateOfferRequest:
      type: object
      properties:
        restaurantId:
          type: integer
        titleEn:
          type: string
        titleAr:
          type: string
        descriptionEn:
          type: string
        descriptionAr:
          type: string
        discountPercent:
          type: number
        originalPrice:
          type: number
        currency:
          type: string
          default: SAR
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        totalCapacity:
          type: integer
      required:
        - restaurantId
        - titleEn
        - titleAr
        - originalPrice
        - validFrom
        - validUntil

    # ─── Vouchers ─────────────────────────────────────────────────────────
    Voucher:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        offerId:
          type: integer
        userId:
          type: integer
        restaurantId:
          type: integer
        restaurantNameEn:
          type: string
        restaurantNameAr:
          type: string
        value:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [active, used, expired]
        validUntil:
          type: string
          format: date-time
        giftMessage:
          type: string
        isGift:
          type: boolean
        redeemedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - restaurantId
        - restaurantNameEn
        - restaurantNameAr
        - value
        - currency
        - status
        - isGift
        - createdAt
    PurchaseVoucherRequest:
      type: object
      properties:
        offerId:
          type: integer
      required:
        - offerId
    GiftVoucherRequest:
      type: object
      properties:
        recipientPhone:
          type: string
        recipientEmail:
          type: string
        giftMessage:
          type: string

    # ─── Reviews ──────────────────────────────────────────────────────────
...
    Event:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        restaurantNameEn:
          type: string
        restaurantNameAr:
          type: string
        titleEn:
          type: string
        titleAr:
          type: string
        descriptionEn:
          type: string
        descriptionAr:
          type: string
        imageUrl:
          type: string
        eventDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        ticketPrice:
          type: number
        currency:
          type: string
        totalCapacity:
          type: integer
        remainingCapacity:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - restaurantId
        - restaurantNameEn
        - restaurantNameAr
        - titleEn
        - titleAr
        - eventDate
        - isActive
        - createdAt
    CreateEventRequest:
      type: object
      properties:
        restaurantId:
          type: integer
        titleEn:
          type: string
        titleAr:
          type: string
        descriptionEn:
          type: string
        descriptionAr:
          type: string
        imageUrl:
          type: string
        eventDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        ticketPrice:
          type: number
        currency:
          type: string
          default: SAR
        totalCapacity:
          type: integer
      required:
        - restaurantId
        - titleEn
        - titleAr
        - eventDate