# lib/api-zod/src/generated/types/purchaseVoucherRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tabaq | طبق - Food & Dining Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseVoucherRequest {
  offerId: number;
}